requirements.txt
cffi==1.1.0
colorama==0.3.3
cryptography==0.9
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.7.7
django-admin-bootstrapped==2.4.0
django-appconf==1.0.1
django-autoslug==1.8.0
django-bootstrap-form==3.2
django-bootstrap3==5.4.0
django-cities-light==3.1.2
django-countries==3.3
django-enumfield==1.2.1
django-extensions==1.5.6
django-extra-views==0.7.0
django-shortuuidfield==0.1.2
django-toolbelt==0.0.1
django-user-accounts==1.1
ecdsa==0.13
enum34==1.0.4
Fabric==1.10.1
gunicorn==19.3.0
idna==2.0
ipaddress==1.0.7
ndg-httpsclient==0.4.0
paramiko==1.15.2
pathlib==1.0.1
Pillow==2.9.0
pinax-theme-bootstrap==5.7.4
psutil==3.2.0
psycopg2==2.6
pyasn1==0.1.7
pycparser==2.13
pycrypto==2.6.1
pyOpenSSL==0.15.1
pytz==2015.4
shortuuid==0.4.2
six==1.9.0
static3==0.5.1
Unidecode==0.4.18
